requirements.txt
fpdf==1.7.2
lxml==5.3.0
python-docx==1.1.2
typing_extensions==4.12.2
xlwt==1.3.0